Expose orderbook main function in lib crate
